Single listener for moving pieces
Listening on Square was causing issues because the boundaries for hover
and click have some overlap so sometimes we would indicate hovering in
one square while clicking on another because the pointer is in both
squares.  Instead, use a single listener on Board and compute the square
using clientX and clientY so the mouse is only hovering over a single
square at a time.