Fix window layout issues
Only measure right pane and bottom left pane once, when the window is
first loaded.  If we measure after resize they may be narrower than
expected because the board on the left is too large.

Call setDimensions in Board every time it is rerendered, so it catches
any possible resizes from the parent component changing size.