Do not show leading zeroes in clock
For hours, just show the hours.  So time controls over 10 hours might
look funny at that transition but who will play with that much time?
For minutes, if <10 minutes are left render a transparent 0 in front of
the number.  This makes it more readable ("9:35" vs "09:35") and the
transparent 0 prevents the number from shifting at 10 minutes.