Maybe better requestanimationframe for clock
Not sure if this does anything, but might prevent multiple calls to
requestAnimationFrame if React calls the function component more than
once per frame.  Depending on the implementation this could also prevent
the number of calls from growing over time since we initialize a single
call to a requestanimationframe loop and cancel it when unmounted.